machines: split actions file
actions.es6 file contains both provider and store actions.
Let's keep them in seperate files to increase readability
and make it easier for new maintainers to navigate through
the package.

Together with the move the functions were also sorted in
aphabetical order.
