+++ pkg/machines/components/diskAdd.jsx
@@ -23,7 +23,7 @@ import DialogPattern from 'cockpit-components-dialog.jsx';
 import Select from "cockpit-components-select.jsx";
 
 import { mouseClick, units, convertToUnit, digitFilter, toFixedPrecision } from '../helpers.es6';
-import { volumeCreateAndAttach, attachDisk, getVm, getStoragePools } from '../actions.es6';
+import { volumeCreateAndAttach, attachDisk, getVm, getStoragePools } from '../actions/provider-actions.es6';
 
 import './diskAdd.css';
