+++ pkg/machines/components/vmLastMessage.jsx
@@ -18,7 +18,7 @@
  */
 import React from "react";
 import { vmId } from '../helpers.es6';
-import { deleteVmMessage } from '../actions.es6';
+import { deleteVmMessage } from '../actions/store-actions.es6';
 import { Alert } from './notification/inlineNotification.jsx';
 
 const VmLastMessage = ({ vm, dispatch }) => {
