+++ pkg/ovirt/index.es6
@@ -20,7 +20,7 @@ import '../lib/polyfills.js'; // once per application
 
 import React from 'react';
 import store from './store.es6';
-import { initDataRetrieval } from '../machines/actions.es6';
+import { initDataRetrieval } from '../machines/actions/provider-actions.es6';
 import { logDebug } from '../machines/helpers.es6';
 
 import Provider from './provider.es6';
